Zaimplementowano zmianę zdjęcia samochodu

+++ src/main/java/com/example/carrentalsystem/Repositories/CarImageRepository.java
@@ -3,7 +3,6 @@ package com.example.carrentalsystem.Repositories;
 import com.example.carrentalsystem.Models.CarImage;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CarImageRepository extends JpaRepository<CarImage, Long> {
+public interface CarImageRepository extends JpaRepository<CarImage, Integer> {
     CarImage getByImageID(Integer imageID);
-
 }
